fix(merge group): remove sorting of skin bind node IDs (#322)
Remove a mistake introduced in an earlier commit, skin bind node ids should of course not be sorted... This will cause the bindings to be mixed up in the final i3d. When sorted the node will be the same, but what its bind to/visually move in GE can potentially be something else.