feat(material): auto-assign default texture slots for materials using vehicleShader (#340)
Materials that use vehicleShader are expected to provide certain base textures for the shader to work properly, but these may be missing in Blender setups. In practice it can be tedious to ensure that every material (especially small or less noticeable parts) has these maps assigned. To reduce errors and make the experience more pleasant for everyone, the exporter now fills in the "expected" default maps automatically when they’re not set:

- $data/shared/white_diffuse.dds → Texture
- $data/shared/default_vmask.dds → Glossmap
- $data/shared/default_normal.dds → Normalmap