fix(properties): enable relative paths for file/directory properties (#356)
This should preferably go in after #331 (since it requires 4.5)

Blender 4.5 introduced the PATH_SUPPORTS_BLEND_RELATIVE option, which is required for StringProp deginitions using FILE_PATH or DIR_PATH to support relative paths. Without this option, relative paths appear as invalid (red) or are set as absolute when using the file browser.

https://projects.blender.org/blender/blender/pulls/137060